https://github.com/mP1/walkingkooka-spreadsheet/pull/7624 compare.provider package extracted
- https://github.com/mP1/walkingkooka-spreadsheet/pull/7624
- compare.provider package extracted
